stage-controller/requirements.txt
numpy
typing-extensions
sangaboard
pyserial